Add files via upload
A simple form program that calculates the sales tax of user input data. The user will input a price and a tax rate. The form will calculate the final total and display the value for the user. Two buttons were added alongside the calculation. One button will clear the text fields and the other will exit the program.