Add files via upload
This program extends from Exercise 3. Exception handling was added and values are formatted to be only three decimal places. The user now submits a value of their choosing in the input field. Two additional buttons were added for the user and the existing button was altered. The user can exit the program with the "Exit" button, the user can press either the "Convert to kilograms" or the "Convert to pounds" button to showcase the proper result. The two labels will change based off of the user input and user button selection.