Add files via upload
Created a Windows Form Application that will allow the user to select a file to input into the program. Statistics are displayed to showcase the file converted to all lower-case, the first and last words alphabetically, the longest word, and the word with the most vowels. All of the statistics are displayed in text boxes within the program and output into a text file.