Add files via upload
Created a windows form application that allows the user to roll a set of dice. The program will continuously roll the dice until double 1's is achieved. The number of sides on the dice is randomly set by the program to be between 4 and 20 sides. Each dice roll is output into the console and labels within the form are updated to showcase the number of sides, the number of rolls, and the final results.